+++ lib/chewy/type/base.rb
@@ -2,8 +2,8 @@ require 'chewy/index/search'
 require 'chewy/type/mapping'
 require 'chewy/type/wrapper'
 require 'chewy/type/observe'
+require 'chewy/type/actions'
 require 'chewy/type/import'
-require 'chewy/type/reset'
 require 'chewy/type/adapter/object'
 require 'chewy/type/adapter/active_record'
 
@@ -14,8 +14,8 @@ module Chewy
       include Mapping
       include Wrapper
       include Observe
+      include Actions
       include Import
-      include Reset
 
       singleton_class.delegate :client, to: :index